[LIVY-1011] Update CI environment, fix PythonInterpreterSpec
**What changes were proposed in this pull request?**

The CI environment broke due to library version conflicts between the
image and the checkout action.

The livy-ci Docker image needs to be updated with a more recent version
of Ubuntu, ensuring that the unit tests all pass. This includes fixing
the issues with PythonInterpreterSpec with Python 3.10+ (which is the
default in Ubuntu 20.04).

**How was this patch tested?**

CI and unit test runs in private fork of the repo.