[LIVY-1011] Update CI environment, fix PythonInterpreterSpec
**What changes were proposed in this pull request?**

The CI environment is broken due to library version conflicts between
the image and the checkout action.

The livy-ci Docker image needs to be updated with a more recent version
of Ubuntu, ensuring that the unit tests and integration tests all pass.

Due to a bug in the GitHub workflows, tests had only been executed with
Spark 2.  Now that this is fixed, we also need to make updates to some
of the Python-based tests to work with Python 3.

Two R-based integration tests are now ignored in the Spark 2 integration
test runs because SparkR 2 does not work with R 4, and the last version
of Ubuntu with R 3.6 is Ubuntu 20. If we want to re-enable these tests,
we may need to build R 3.6 for Ubuntu 24.

**How was this patch tested?**

CI and unit test runs in a private fork of the repo.